requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
altair==5.0.1
asgiref==3.7.2
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
Automat==22.10.0
backcall==0.2.0
beautifulsoup4==4.12.2
bert-extractive-summarizer==0.10.1
blinker==1.6.2
blis==0.7.9
branca==0.6.0
breadability==0.1.20
cachetools==5.3.1
catalogue==2.0.8
certifi==2023.5.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
cmake==3.26.4
confection==0.1.0
constantly==15.1.0
contourpy==1.1.0
cryptography==41.0.1
cssselect==1.2.0
cycler==0.11.0
cymem==2.0.7
Cython==0.29.35
dateparser==1.1.8
decorator==5.1.1
directory-tree==0.0.3.1
Django==4.2.2
docopt==0.6.2
enum-compat==0.0.3
executing==1.2.0
Faker==18.11.2
favicon==0.7.0
filelock==3.12.2
folium==0.14.0
fonttools==4.40.0
frozenlist==1.3.3
fsspec==2023.6.0
funcy==2.0
gensim==4.3.1
gitdb==4.0.10
GitPython==3.1.31
google-api-core==2.11.1
google-api-python-client==2.92.0
google-auth==2.21.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.59.1
greenlet==2.0.2
htbuilder==0.6.1
httplib2==0.22.0
huggingface-hub==0.15.1
hyperlink==21.0.0
idna==3.4
importlib-metadata==6.7.0
importlib-resources==5.12.0
incremental==22.10.0
install==1.3.5
instaloader==4.9.6
ipython==8.14.0
itemadapter==0.8.0
itemloaders==1.1.0
jedi==0.18.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.1
jsonpickle==3.0.1
jsonschema==4.17.3
kiwisolver==1.4.4
langcodes==3.3.0
lit==16.0.6
lxml==4.9.2
Markdown==3.4.3
markdown-it-py==3.0.0
markdownlit==0.0.7
MarkupSafe==2.1.3
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdurl==0.1.2
more-itertools==9.1.0
mpmath==1.3.0
multidict==6.0.4
murmurhash==1.0.9
networkx==3.1
nltk==3.8.1
numexpr==2.8.4
numpy==1.25.0
nvidia-curand-cu11==10.2.10.91
oauthlib==3.2.2
openai==0.27.8
packaging==23.1
pandas==2.0.3
parsel==1.8.1
parso==0.8.3
pathy==0.10.2
patsy==0.5.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
pipdeptree==2.9.3
playwright==1.35.0
plotly==5.15.0
plotly-express==0.4.1
preshed==3.0.8
prompt-toolkit==3.0.38
Protego==0.2.1
protobuf==4.23.3
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==12.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycountry==22.3.5
pycparser==2.21
pydantic==1.10.10
pydeck==0.8.1b0
PyDispatcher==2.0.7
pyee==9.0.4
Pygments==2.15.1
pyLDAvis==3.4.1
pymdown-extensions==10.0.1
Pympler==1.0.1
pyOpenSSL==23.2.0
pyparsing==3.1.0
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
pyvis==0.3.2
PyYAML==6.0
queuelib==1.6.2
regex==2023.6.3
requests==2.31.0
requests-file==1.5.1
requests-oauthlib==1.3.1
rich==13.4.2
rsa==4.9
sacremoses==0.0.53
safetensors==0.3.1
Sastrawi==1.0.1
scikit-learn==1.2.2
scipy==1.11.1
Scrapy==2.9.0
seaborn==0.12.2
sentencepiece==0.1.99
service-identity==23.1.0
six==1.16.0
smart-open==6.3.0
smmap==5.0.0
sonopy==0.1.2
soupsieve==2.4.1
spacy==3.5.4
spacy-legacy==3.0.12
spacy-loggers==1.0.4
sqlparse==0.4.4
srsly==2.4.6
st-annotated-text==4.0.0
stack-data==0.6.2
statsmodels==0.14.0
streamlit==1.24.0
streamlit-camera-input-live==0.2.0
streamlit-card==0.0.61
streamlit-embedcode==0.1.2
streamlit-extras==0.2.7
streamlit-faker==0.0.2
streamlit-image-coordinates==0.1.5
streamlit-keyup==0.2.0
streamlit-tags==1.2.8
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==1.0.2
sympy==1.12
tenacity==8.2.2
thinc==8.1.10
threadpoolctl==3.1.0
TikTokApi==5.2.2
tldextract==3.4.4
tokenizers==0.13.3
toml==0.10.2
toolz==0.12.0
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
transformers==4.30.2
tweepy==4.14.0
Twisted==22.10.0
typer==0.9.0
typing-extensions==4.7.0
tzdata==2023.3
tzlocal==4.3.1
uritemplate==4.1.1
urllib3==1.26.16
vader==0.0.3
vaderSentiment==3.3.2
validators==0.20.0
w3lib==2.1.1
wasabi==1.1.2
watchdog==3.0.0
wcwidth==0.2.6
yarl==1.9.2
zipp==3.15.0
zope.interface==6.0
